Reimplement the morphic layout of our SearchInputField,  Fixes #1687
The implementation is a bit cursed. Manually creates a Spec overlay in morphic and proceeds to properly place the "clear" button on the top right corner.
There was a comment saying that the original author preferred the button to be centered. This is easy now but I disagree with the original author.

Co-authored-by: Matija Kljajic <matijakljajic173@gmail.com>